+++ cg/meta/orders/schema.py
@@ -247,7 +247,7 @@ FASTQ_SAMPLE = {
 }
 
 RML_SAMPLE = {
-    # 1604:9 Orderform Ready made libraries (RML)
+    # 1604:10 Orderform Ready made libraries (RML)
     # Order portal specific
     "priority": str,
     # "This information is required"
@@ -258,6 +258,7 @@ RML_SAMPLE = {
     "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "volume": str,
     "concentration": str,
+    "concentration_sample": OptionalNone(TypeValidatorNone(str)),
     "index": str,
     "index_number": OptionalNone(TypeValidatorNone(str)),  # optional for NoIndex
     # "Required if Plate"
@@ -267,7 +268,6 @@ RML_SAMPLE = {
     "index_sequence": OptionalNone(TypeValidatorNone(str)),
     # "Not required"
     "comment": OptionalNone(TypeValidatorNone(str)),
-    "capture_kit": OptionalNone(validators.Any(CAPTUREKIT_OPTIONS)),
 }
 
 MICROSALT_SAMPLE = {
@@ -290,7 +290,7 @@ MICROSALT_SAMPLE = {
     # "Required if "Other" is chosen in column "Species""
     "organism_other": OptionalNone(TypeValidatorNone(str)),
     # "These fields are not required"
-    "concentration_weight": OptionalNone(TypeValidatorNone(str)),
+    "concentration_sample": OptionalNone(TypeValidatorNone(str)),
     "quantity": OptionalNone(TypeValidatorNone(str)),
     "comment": OptionalNone(TypeValidatorNone(str)),
 }
@@ -311,7 +311,7 @@ METAGENOME_SAMPLE = {
     "container_name": OptionalNone(TypeValidatorNone(str)),
     "well_position": OptionalNone(TypeValidatorNone(str)),
     # "This information is not required"
-    "concentration_weight": OptionalNone(TypeValidatorNone(str)),
+    "concentration_sample": OptionalNone(TypeValidatorNone(str)),
     "quantity": OptionalNone(TypeValidatorNone(str)),
     "extraction_method": OptionalNone(TypeValidatorNone(str)),
     "comment": OptionalNone(TypeValidatorNone(str)),