+++ tests/conftest.py
@@ -3,6 +3,7 @@
 """
 import copy
 import datetime as dt
+import json
 import logging
 import shutil
 import os
@@ -299,7 +300,7 @@ def microbial_orderform(orderforms: Path) -> str:
 @pytest.fixture
 def rml_orderform():
     """Orderform fixture for RML samples"""
-    return "tests/fixtures/orderforms/1604.9.rml.xlsx"
+    return "tests/fixtures/orderforms/1604.10.rml.xlsx"
 
 
 @pytest.fixture(name="madeline_output")
@@ -307,6 +308,54 @@ def fixture_madeline_output(apps_dir: Path) -> str:
     """File with madeline output"""
     _file = apps_dir / "madeline/madeline.xml"
     return str(_file)
+
+
+@pytest.fixture
+def mip_order_to_submit() -> dict:
+    """Load an example scout order."""
+    return json.load(open("tests/fixtures/orders/mip.json"))
+
+
+@pytest.fixture
+def mip_rna_order_to_submit() -> dict:
+    """Load an example rna order."""
+    return json.load(open("tests/fixtures/orders/mip_rna.json"))
+
+
+@pytest.fixture
+def external_order_to_submit() -> dict:
+    """Load an example external order."""
+    return json.load(open("tests/fixtures/orders/external.json"))
+
+
+@pytest.fixture
+def fastq_order_to_submit() -> dict:
+    """Load an example fastq order."""
+    return json.load(open("tests/fixtures/orders/fastq.json"))
+
+
+@pytest.fixture
+def rml_order_to_submit() -> dict:
+    """Load an example rml order."""
+    return json.load(open("tests/fixtures/orders/rml.json"))
+
+
+@pytest.fixture
+def metagenome_order_to_submit() -> dict:
+    """Load an example metagenome order."""
+    return json.load(open("tests/fixtures/orders/metagenome.json"))
+
+
+@pytest.fixture
+def microbial_order_to_submit() -> dict:
+    """Load an example microbial order."""
+    return json.load(open("tests/fixtures/orders/microsalt.json"))
+
+
+@pytest.fixture
+def balsamic_order_to_submit() -> dict:
+    """Load an example cancer order."""
+    return json.load(open("tests/fixtures/orders/balsamic.json"))
 
 
 # Compression fixtures
@@ -754,6 +803,7 @@ def fixture_base_store(store) -> Store:
             sequencing_depth=0,
             is_external=True,
             percent_kth=80,
+            percent_reads_guaranteed=75,
             target_reads=10,
         ),
         store.add_application(
@@ -763,6 +813,7 @@ def fixture_base_store(store) -> Store:
             sequencing_depth=0,
             is_external=True,
             percent_kth=80,
+            percent_reads_guaranteed=75,
             target_reads=10,
         ),
         store.add_application(
@@ -772,6 +823,7 @@ def fixture_base_store(store) -> Store:
             sequencing_depth=30,
             accredited=True,
             percent_kth=80,
+            percent_reads_guaranteed=75,
             target_reads=10,
         ),
         store.add_application(
@@ -780,6 +832,7 @@ def fixture_base_store(store) -> Store:
             description="Ready-made",
             sequencing_depth=0,
             percent_kth=80,
+            percent_reads_guaranteed=75,
             target_reads=10,
         ),
         store.add_application(
@@ -791,6 +844,7 @@ def fixture_base_store(store) -> Store:
             target_reads=30,
             limitations="some",
             percent_kth=80,
+            percent_reads_guaranteed=75,
         ),
         store.add_application(
             tag="METLIFR020",
@@ -799,6 +853,7 @@ def fixture_base_store(store) -> Store:
             sequencing_depth=0,
             target_reads=400000,
             percent_kth=80,
+            percent_reads_guaranteed=75,
         ),
         store.add_application(
             tag="METNXTR020",
@@ -807,6 +862,7 @@ def fixture_base_store(store) -> Store:
             sequencing_depth=0,
             target_reads=200000,
             percent_kth=80,
+            percent_reads_guaranteed=75,
         ),
         store.add_application(
             tag="MWRNXTR003",
@@ -814,6 +870,7 @@ def fixture_base_store(store) -> Store:
             description="Microbial whole genome ",
             sequencing_depth=0,
             percent_kth=80,
+            percent_reads_guaranteed=75,
             target_reads=10,
         ),
         store.add_application(
@@ -821,6 +878,7 @@ def fixture_base_store(store) -> Store:
             category="tgs",
             description="RNA seq, poly-A based priming",
             percent_kth=80,
+            percent_reads_guaranteed=75,
             sequencing_depth=25,
             accredited=True,
             target_reads=10,