+++ CHANGELOG.md
@@ -12,15 +12,102 @@ Try to use the following format:
 
 ### Added
 
-### Changed
-
 ### Fixed
+
+### Changed
 
 ## [NG.NG.NG]
 
 ### Added
 
 - Stop on bad values for analysis (pipeline) when adding family through CLI
+
+## 16.12.1
+
+### Fixed
+- Save in Family and Analysis Admin views
+
+## 16.12.0
+
+### Added
+
+- Added support for RML-orderforms as json-file
+
+## [16.11.1]
+
+### Fixed
+
+- Lock dependency for xlrd so that we can parse modern excel files
+
+## 16.11.0
+
+### Added
+- cg set family [CASEID] now has the option --data-delivery 
+
+## [16.10.4]
+
+### Fixed
+- Bug when building tag definition for balsamic-analysis delivery
+
+## [16.10.3]
+
+### Fixed
+- Use `--json` when exporting causative variants from scout
+
+## [16.10.2]
+
+### Fixed
+- Use correct pipeline name when cleaning mip analysis dirs
+
+## [16.10.1]
+
+### Added
+
+- Adds new mip-dna tags to hk
+
+## [16.10.0]
+
+### Added
+
+- Adds new delivery type to balsamic-analysis
+
+## [16.9.0]
+
+### Added
+
+- column percent_reads_guaranteed to application table
+
+## [16.8.1]
+
+### Fixed
+
+- Bug in the interactions with Scout when cleaning Scout cases
+- Bug in the interaction with scout in command export_causatives
+
+## [16.8.0]
+
+### Added
+
+- Adds adding samplesheet to HK flowcell bundle to cg transfer flowcell
+
+## [16.7.1]
+
+### Fixed
+
+- Mutacc looks for 'id' instead of '_id' in case export
+- Convert 'other' to '0' for sex in case export
+
+## 16.7.0
+
+### Added
+- Show sample priorities in created ticket
+
+## [16.6.0]
+
+### Changed
+
+- Split generating config into its own command
+- Delete old load config when running `cg upload scout --re-upload`
 
 ## [16.5.0]